+++ app/controllers/questions_controller.rb
@@ -14,7 +14,7 @@ class QuestionsController < ApplicationController
   def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
       redirect_to @question, notice: 'Your question successfully created.'
@@ -32,8 +32,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.questions.include?(question)
+      question.destroy
+      redirect_to questions_path, notice: "Question was deleted successful"
+    else
+      redirect_to question_path(question), alert: "You don't have permission"
+    end
   end
 
   private
